try fixing attention mask issues
